Search only through meetings with a valid link
Before, this would take all calendar meetings from now onwards and open
the one at the given index.
Now, it will first narrow down the search to meetings with a valid Zoom
link and only then use the given index to make the selection.
The search is limited to a maximum amount of 20 entries total, retrieved
in batches of 5 from Google Calendar.